requirements.txt
fastapi
uvicorn
ollama
python-multipart
sqlmodel
python-dotenv
pdf2image
pytesseract
Pillow
tenacity
